+++ packages/frontend/src/features/semanticViewer/store/semanticViewerSlice.ts
@@ -4,7 +4,7 @@ import {
     DimensionType,
     FieldType,
     SemanticLayerFieldType,
-    type ResultRow,
+    type RawResultRow,
     type SemanticLayerClientInfo,
     type SemanticLayerField,
     type SemanticLayerSortBy,
@@ -111,7 +111,7 @@ function getDimensionTypeFromSemanticLayerFieldType(
 }
 
 export type ResultsAndColumns = {
-    results: ResultRow[];
+    results: RawResultRow[];
     columns: VizSqlColumn[];
     sortBy: [];
 };
@@ -134,7 +134,7 @@ export interface SemanticViewerState {
 
     resultsTableConfig: VizTableConfig | undefined;
 
-    results: ResultRow[];
+    results: RawResultRow[];
     columns: VizSqlColumn[];
 
     selectedDimensions: Record<string, SemanticLayerStateDimension>;
@@ -194,7 +194,7 @@ export const semanticViewerSlice = createSlice({
         setResults: (
             state,
             action: PayloadAction<{
-                results: ResultRow[];
+                results: RawResultRow[];
                 columns: VizSqlColumn[];
             }>,
         ) => {