+++ packages/frontend/src/features/sqlRunner/hooks/useSqlChartResults.ts
@@ -3,7 +3,7 @@ import {
     isErrorDetails,
     type ApiError,
     type ApiJobScheduledResponse,
-    type ResultRow,
+    type RawResultRow,
 } from '@lightdash/common';
 import { useQuery } from '@tanstack/react-query';
 import { lightdashApi } from '../../../api';
@@ -31,7 +31,7 @@ const getSqlChartResults = async ({
         !isErrorDetails(job.details)
             ? job.details.fileUrl
             : undefined;
-    const results = await getResultsFromStream<ResultRow>(url);
+    const results = await getResultsFromStream<RawResultRow>(url);
 
     return {
         results,